Изменение ViewModel | Добавлена работа с новыми интерфейсами взамен старых.
